Fix various GC-related misbehaviour.
The data section of the callspec custom block now stores a pointer to the cif,
not the cif itself, since ffi_closures store the cif address and the GC can
move the block contents.

The callback_handler function now unregisters local roots.

The boxedfn structure now holds a reference to the callspec to ensure that
it's not collected prematurely.