Don't install the cmx files for ctypes-foreign.
We install two archives (.cmxa files) for ctypes-foreign and pick the
right one at link time using the findlib mt predicate, but we only
install one set of native objects (.cmx files).  Since it's not
possible to have a single set of native objects that's compatible with
both archives we end up with link-time incompatibilities.