work around missing math functions on bsd platforms
This will fix #502 and a related problem on netbsd. I've intentionally
not implemented any kind of feature detection. It's tricky. The header
definitions are available and the functions are understood by gcc/clang,
although they are not implemented by the c standard library. But gcc will
generate opcode for them under certain conditions...