Make Ctypes_ptr.Raw independent from Signed.S
Since `Signed.S` is defined in `integers` (now separate), additions to
this type signature can break `Ctypes_ptr.Raw`.

However, this module is not public API, so it is possible to remove that
module type signature, and even to restrict what is exported to the bare
minimum: `null` and `nativeint` conversion functions.