Select ctypes primitives at runtime
Executing a program that uses `ocaml-ctypes` on a different platform from where
the OCaml code was compiled can result in memory errors due to the wrong pointer
arithmetic. For example, when a program is compiled on a 64-bit machine to OCaml
bytecode and the bytecode is subsequently compiled to 32-bit WebAssembly (Wasm)
code, `ocaml-types` uses 64-bit pointer arithmetic on a 32-bit architecture.

This commit makes the C data model selectable using an environment variable,
i.e., `CTYPES_DATA_MODEL=ILP32` for 32-bit Wasm. The data model can be selected
during compilation or during runtime.